[add]: added new obs objects and existing moved to acceptance/obs
