+++ sbercloud/provider.go
@@ -221,6 +221,7 @@ func Provider() *schema.Provider {
 			"sbercloud_nat_gateway":            nat.DataSourcePublicGateway(),
 			"sbercloud_networking_port":        vpc.DataSourceNetworkingPortV2(),
 			"sbercloud_networking_secgroup":    huaweicloud.DataSourceNetworkingSecGroup(),
+			"sbercloud_obs_buckets":            obs.DataSourceObsBuckets(),
 			"sbercloud_obs_bucket_object":      obs.DataSourceObsBucketObject(),
 			"sbercloud_rds_flavors":            rds.DataSourceRdsFlavor(),
 			"sbercloud_sfs_file_system":        huaweicloud.DataSourceSFSFileSystemV2(),
@@ -320,6 +321,7 @@ func Provider() *schema.Provider {
 			"sbercloud_obs_bucket":                      obs.ResourceObsBucket(),
 			"sbercloud_obs_bucket_object":               obs.ResourceObsBucketObject(),
 			"sbercloud_obs_bucket_policy":               obs.ResourceObsBucketPolicy(),
+			"sbercloud_obs_bucket_acl":                  obs.ResourceOBSBucketAcl(),
 			"sbercloud_rds_instance":                    rds.ResourceRdsInstance(),
 			"sbercloud_rds_parametergroup":              rds.ResourceRdsConfiguration(),
 			"sbercloud_rds_read_replica_instance":       rds.ResourceRdsReadReplicaInstance(),
